Attribute Macro server_fn::default::server

#[server]

Declares that a function is a server function. This means that its body will only run on the server, i.e., when the ssr feature is enabled.

You can specify one, two, three, or four arguments to the server function:

Required: A type name that will be used to identify and register the server function (e.g., MyServerFn).

Optional: A URL prefix at which the function will be mounted when it’s registered (e.g., "/api"). Defaults to "/".

Optional: The encoding for the server function ("Url", "Cbor", "GetJson", or "GetCbor“. See Server Function Encodings below.)

Optional: A specific endpoint path to be used in the URL. (By default, a unique path will be generated.)

// will generate a server function at `/api-prefix/hello`
#[server(MyServerFnType, "/api-prefix", "Url", "hello")]

The server function itself can take any number of arguments, each of which should be serializable and deserializable with serde.

#[server(ReadPosts, "/api")]
pub async fn read_posts(how_many: u8, query: String) -> Result<Vec<Post>, ServerFnError> {
  // do some work on the server to access the database
  todo!()
}

Note the following:

Server functions must be async. Even if the work being done inside the function body can run synchronously on the server, from the client’s perspective it involves an asynchronous function call.

Server functions must return Result<T, ServerFnError>. Even if the work being done inside the function body can’t fail, the processes of serialization/deserialization and the network call are fallible.

Return types must implement Serialize. This should be fairly obvious: we have to serialize arguments to send them to the server, and we need to deserialize the result to return it to the client.

Arguments must be implement Serialize and DeserializeOwned. They are serialized as an application/x-www-form-urlencoded form data using serde_qs or as application/cbor using cbor.